Better shader chunking and delayed decompression
Shader variants were previously put into chunks randomly and CompiledShaderGroup::FromFile decompressed every shader in the .csg file. I changed this behaviour: now the value of the first few options determine the compression block of the variant. This way only the block of the actually used shaders are decompressed at runtime. The block is cached, so similar variants can be later accessed faster.